Update 코드 정리
