+++ Projects/App/Sources/Present/Detail/ViewController/Modal/Model/OptionData.swift
@@ -4,10 +4,4 @@ struct OptionData {
     let image: UIImage
     let text: String
     let color: UIColor
-    
-    init(image: UIImage, text: String, color: UIColor) {
-        self.image = image 
-        self.text = text
-        self.color = color
-    }
 }
